issue with writing to file
When retesting the code we noticed that the files were not created properly. 

I think that it may have been due to the way I was creating them. As such that was changed to the original way (prior to code release). 
As for the status check, I changed it that instead of checking each time whether or not files and directories were created, it only checks at the very end by calling "create_file_check()".